requirements.txt
groq==0.4.2
python-dotenv==1.0.0
streamlit==1.31.0
Pillow==10.0.0
